theming should live throgh closing app successfully now

+++ app/src/main/java/com/example/wildfire_fixed_imports/model/DataClasses.kt
@@ -156,7 +156,7 @@ data class WebBELocation(
         val longitude: Double,
         val notification_timer: Int,
         val notifications: Boolean,
-        val radius: Int,
+        val radius: Double,
         val user_id: Int
 
 ) {
@@ -176,7 +176,7 @@ data class WebBELocation(
             val longitude: Double,
             val notification_timer: Int,
             val notifications: Boolean,
-            val radius: Int
+            val radius: Double
             // val user_id: Int
     )
 }
